Add an 'I' Field
Some fields in the JSON are numbers -- and numbers in JSON are all one type. Aeson sees them as this "Scientific" number -- a float or double? So if WordPress has a post with an "id" 34234, then Offset fills it as "34234.0". Which is bad when we're creating links to "author/${wpId}".

The problem seems to be that Fields have some different types, like objects or lists, but there's no "type" that says anything about the content of a field.

Ideally, it would be getting a JSON Number, then trying to convert that into an Integer. But what the "build splices"