Users can specify a date format for dates
I changed the "format" attribute name to "wp_format", because it specifies the format you expect the WordPress date to be in. Since WordPress just has a default date type, I also made this an optional field.

Most of the time, we want days to be zero-padded to two characters, so I made that the default.

I also made the formats of each date field configurable using the formatDate function from the same library as our date parsing function. This allows users to do stuff like show the name of the month really easily!! It's just another format string.